yp (#12)
* Ensure depth test is configured when rendering clouds

The state of the depth test prior to cloud rendering is
undefined. After rendering, it is expected to be
disabled again.

* Fix rounding error in ColorMixer#mix

Rounding of the values now happens after the 16-bit
intermediaries are added together.

This affected some animated textures, causing them to
exhibit flickering behavior.

* Add additional optimized block models

This covers the following additional blocks:
- Cauldrons
- Brewing Stands
- Bells

Co-authored-by: JellySquid <jellysquid@pm.me>

* Avoid marking the section graph as dirty if state didn't change (#2886)

Avoids rebuilding the render lists and doing a graph search
more often than necessary by checking if the section actually
changed in a way that's relevant to the graph search.

For worlds that update their blocks frequently (every tick or
every redstone tick) this avoids half the graph searches. Some
graph searches are still necessary to schedule rebuild tasks,
but when the task results come back, this doesn't do another
graph search unless the section's visibility data or build state
changed in a way that needs the render list to be updated.

* Use larger bounding box for nearby sections in frustum check (#2879)

This fixes some problems where very large block entities in
nearby sections may be incorrectly culled. But it does not
comprehensively fix the problem for all other sections,
since that would require visiting the 27-neighborhood of
every section, which is too slow.

* Bump version to 0.6.1

* Bump dependency versions

* Update mod manifest

* Ensure ItemRenderContext.isDefaultTranslucent is initialized

* Update compatible mods listing

* Update mod manifest to restrict Minecraft versions

* Update to Minecraft 1.21.4

* Update NeoForge manifest for Minecraft 1.21.4

* Fix glyph effect orientation

* Fix hidden surface elimination in fluid rendering for waterlogged blocks (#2907)

* Fix detection for specific Intel OpenGL ICDs

The OpenGL ICD name now includes the file extension,
which the regex expressions were not matching.

* Avoid showing the incompatible driver error in some cases

For systems with hybrid graphics, it may be the case
that an incompatible graphics driver is installed, but that
it isn't used for the OpenGL context.

We can avoid showing errors in this situation by checking
the vendor string of the context immediately after
creation.

This is not the most robust check, but in practice, a single
system should not have multiple graphics drivers installed
from the same vendor, so checking the string should be
relatively safe.

* Fix lambda mappings

* Bump version and dependency requirements

* Use correct coordinates for sorting chunk sections (#2924)

Fix section and region sorting by using the correct section coordinate instead of the integer part of the camera transform, which is incorrect near the origin.

Closes #2918

* Update README.md

* Update textures for Leaves variants

* Added "Pale Oak Leaves" for Minecraft 1.21.4.
* Reduced the file size of all block textures.

* Clean up BitArray class

* Use link.caffeinemc.net domain for some URLs

* Use ShellExecuteW from message box callbacks

This fixes a regression caused by 26f4263.

The underlying problem is that accessing Java's AWT *after*
LWJGL3 has initialized is not possible.

Minecraft has a utility class which uses rundll32 internally,
but we cannot access that due to classloader restrictions on
NeoForge.

That leaves us with having to implement the call ourselves,
and simply using Shell32 directly (like we do for other
Windows APIs) seems easiest.

* Do not bake ambient lighting into cached per-face light data

Fixes #2806

* Switch to Parchment mappings 2024.12.07

* Bump version and dependencies

* Do not cache ambient brightness at initialization

The world may not be assigned to the renderer at
initialization, which is the case for non-terrain
rendering (i.e. block entities.)

Likely, there is no performance benefit to caching
this data in the first place, so the easiest solution
is to just remove the code.

* Bump version to 0.6.5

* Fix transforms not being run on fast path

* Bump version to 0.6.6

* Rename model texture references to match Vanilla (#2958)

Custom models which extended these base models would not properly have their textures applied, as the texture references were accidentally changed.

* Do not apply optimizations to sprites with special tickers

* Aggressively optimize entity rendering

This is anywhere from 10 to 15% faster depending
on what entities are being rendered.

Most of the improvements come from more efficiently
laying out the cuboid data and coalescing neighboring
32-bit values into 64-bit words.

Furthermore, vertex positions are calculated by
extracting vectors from the pose matrix and adding
them to the origin vertex, which avoids many
matrix multiplications.

Co-authored-by: MoePus <547007249@qq.com>

* Avoid quaternion transforms in particle rendering

The billboard geometry can be computed using the
camera's left and up vectors, saving some cycles.

When rendering thousands of billboard particles, this
was ~10% faster than baseline in my observation.

Co-authored-by: MoePus <547007249@qq.com>

* Add shader source line annotations (#2691)

---------

Co-authored-by: JellySquid <jellysquid@pm.me>
Co-authored-by: muzikbike <52297970+muzikbike@users.noreply.github.com>
Co-authored-by: douira <douira@users.noreply.github.com>
Co-authored-by: IMS212 <ims@irisshaders.dev>
Co-authored-by: MoePus <547007249@qq.com>